move eigen/reduce.h imple into cpu/reduce.h
