+++ paddle/pten/kernels/cpu/math_kernel.cc
@@ -21,6 +21,7 @@
 #include "paddle/pten/kernels/cpu/elementwise_impl.h"
 #include "paddle/pten/kernels/cpu/reduce.h"
 #include "paddle/pten/kernels/funcs/elementwise_functor.h"
+#include "paddle/pten/kernels/funcs/reduce_functor.h"
 
 // See Note [ Why still include the fluid headers? ]
 #include "paddle/fluid/framework/eigen.h"
@@ -54,19 +55,19 @@ namespace pten {
   }
 
 template <typename T, typename Context>
-void MeanKernel(const Context& dev_ctx,
+void MeanKernel(const Context& ctx,
                 const DenseTensor& x,
                 const std::vector<int64_t>& dims,
                 bool keep_dim,
                 bool reduce_all,
                 DenseTensor* out) {
   auto out_dtype = x.dtype();
-  pten::Reduce<CPUContext, T, pten::eigen::MeanFunctor>(
-      dev_ctx, x, reduce_all, dims, keep_dim, out_dtype, out);
+  pten::Reduce<CPUContext, T, pten::funcs::MeanFunctor>(
+      ctx, x, reduce_all, dims, keep_dim, out_dtype, out);
 }
 
 template <typename T, typename Context>
-void DivideKernel(const Context& dev_ctx,
+void DivideKernel(const Context& ctx,
                   const DenseTensor& x,
                   const DenseTensor& y,
                   int axis,
@@ -75,30 +76,30 @@ void DivideKernel(const Context& dev_ctx,
   out->mutable_data<T>();
   if (x.dims() == y.dims() && std::is_floating_point<T>::value) {
     SameDimsElementwiseCompute<SameDimsDivideFunctor<CPUContext, T>>()(
-        dev_ctx, x, y, out);
+        ctx, x, y, out);
   } else {
     auto x_dims = x.dims();
     auto y_dims = y.dims();
     if (x_dims.size() >= y_dims.size()) {
       ElementwiseCompute<funcs::DivideFunctor<T>, T>(
-          dev_ctx, x, y, axis, funcs::DivideFunctor<T>(), out);
+          ctx, x, y, axis, funcs::DivideFunctor<T>(), out);
     } else {
       ElementwiseCompute<funcs::InverseDivideFunctor<T>, T>(
-          dev_ctx, x, y, axis, funcs::InverseDivideFunctor<T>(), out);
+          ctx, x, y, axis, funcs::InverseDivideFunctor<T>(), out);
     }
   }
 }
 
 template <typename T, typename Context>
-void SumKernel(const Context& dev_ctx,
+void SumKernel(const Context& ctx,
                const DenseTensor& x,
                const std::vector<int64_t>& dims,
                bool keep_dim,
                bool reduce_all,
                DataType out_dtype,
                DenseTensor* out) {
-  pten::Reduce<CPUContext, T, pten::eigen::SumFunctor>(
-      dev_ctx, x, reduce_all, dims, keep_dim, out_dtype, out);
+  pten::Reduce<CPUContext, T, pten::funcs::SumFunctor>(
+      ctx, x, reduce_all, dims, keep_dim, out_dtype, out);
 }
 
 // Create the definition of Add
