+++ paddle/pten/kernels/math_kernel.h
@@ -17,11 +17,12 @@ limitations under the License. */
 #include "paddle/pten/api/lib/utils/storage.h"
 #include "paddle/pten/core/dense_tensor.h"
 #include "paddle/pten/include/infermeta.h"
+#include "paddle/pten/kernels/empty_kernel.h"
 
 namespace pten {
 
 template <typename T, typename Context>
-void MeanKernel(const Context& dev_ctx,
+void MeanKernel(const Context& ctx,
                 const DenseTensor& x,
                 const std::vector<int64_t>& dims,
                 bool keep_dim,
@@ -29,7 +30,7 @@ void MeanKernel(const Context& dev_ctx,
                 DenseTensor* out);
 
 template <typename T, typename Context>
-void AddKernel(const Context& dev_ctx,
+void AddKernel(const Context& ctx,
                const DenseTensor& x,
                const DenseTensor& y,
                int axis,
@@ -118,6 +119,36 @@ DenseTensor Multiply(const ContextT& dev_ctx,
           dev_ctx.GetPlace()),
       std::move(out_meta));
   MultiplyKernel<T, ContextT>(dev_ctx, x, y, axis, &dense_out);
+  return dense_out;
+}
+
+template <typename T, typename Context>
+DenseTensor Mean(const Context& ctx,
+                 const DenseTensor& x,
+                 const std::vector<int64_t>& axis,
+                 bool keep_dim) {
+  auto out_meta = ReduceInferMeta(x.meta(), axis, keep_dim);
+  auto dense_out = pten::Empty<T, Context>(ctx, std::move(out_meta));
+  bool reduce_all = false;
+  MeanKernel<T, Context>(ctx, x, axis, keep_dim, reduce_all, &dense_out);
+  return dense_out;
+}
+
+template <typename T, typename Context>
+DenseTensor Sum(const Context& ctx,
+                const DenseTensor& x,
+                const std::vector<int64_t>& axis,
+                DataType dtype,
+                bool keep_dim) {
+  auto out_meta = ReduceInferMeta(x.meta(), axis, keep_dim, dtype);
+  auto dense_out = pten::Empty<T, Context>(ctx, std::move(out_meta));
+
+  // The real value of reduce_all will be get in kernel
+  // so use default value(false) is OK.
+  bool reduce_all = false;
+
+  SumKernel<T, Context>(
+      ctx, x, axis, keep_dim, reduce_all, out_meta.dtype, &dense_out);
   return dense_out;
 }
